+++ Toggly.FeatureManagement.Next/nextjs-toggly-core/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.11.1
+
+2026-09-17
+
+### Fixed
+- Default usage/metrics host is `https://metrics.toggly.io/`.
+
 ## 1.11.0
 
 2026-09-12