+++ Toggly.FeatureManagement.Node/toggly-node-core/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.9.2
+
+2026-09-17
+
+### Fixed
+- Default usage/metrics gRPC host is `https://metrics.toggly.io/`.
+
 ## 0.9.1
 
 2026-09-12